Updated gmusic api version
Even though the version of the gmusic api is specified I was getting the
new version. There is a slight change to the api that was crashing the
plugin. I fixed the function call to gmusicapi.get_all_playlist_ids.

Updated the plugin version number to reflect the changed pre-requisite.